ProtoConnection should be done.
Flow control added. I don't think it makes a lot of sense here, since it
adds a loop of sending to the update function.
Added default parameter to send for flags; this controls the priority of
packets.
High priority packets are sent first with flow control, and are
automatically resent if they aren't ackd.